feat(jobs): redesign — overview grid + per-kind pages + history page
Splits the single Jobs page into three surfaces that match how the
surface actually reads:

- /admin/settings/jobs — overview grid. One card per schedulable
  kind with display name, description, enabled pill, cron, and
  last-fired timestamp. Click any card to drill into settings.
- /admin/settings/jobs/:kind — per-kind settings. Renders the
  single-row schedule editor (JobSchedulesSection now accepts a
  kind prop for filtering), the kind-specific config block (for
  ai_suggestions: the existing AISuggestionsJobCard), a Run Now
  button where a manual trigger exists, and a Recent runs panel
  pulled from the unified history filtered by kind.
- /admin/settings/jobs/history — the cross-kind history table
  extracted from the old JobsPage. Breadcrumb back to Jobs.

Non-schedulable kinds (import, enrichment) don't show on the
overview — they surface in history only; their trigger points live
elsewhere in the UI.

JobSchedulesSection gains an optional kind prop for client-side
filtering.

Bundle growth vs the prior PR is marginal (+1 KB gzipped). The
react-js-cron / antd footprint still dominates.
